Rename incrementVersion to increment-version
